Deleted domino scripts, added setup script and fixed errors

+++ transform_data.py
@@ -46,10 +46,10 @@ def resize_images_to_uniform_size_on_one_axis(images):
 def pad_images_to_be_uniformm_size(images):
     return (pad(image,
                 ((0, 0),
-                 (ceil((96 - image.shape[1]) / 2.0),
-                  floor((96 - image.shape[1]) / 2.0)),
-                 (ceil((96 - image.shape[2]) / 2.0),
-                  floor((96 - image.shape[2]) / 2.0))),
+                 (int(ceil((96 - image.shape[1]) / 2.0)),
+                  int(floor((96 - image.shape[1]) / 2.0))),
+                 (int(ceil((96 - image.shape[2]) / 2.0)),
+                  int(floor((96 - image.shape[2]) / 2.0)))),
                 mode="edge") for image in images)
